Add files via upload
I have uploaded the traffic videos on which we can test the modal. For now we are using YOLOv8n (Nano version of YOLO) for object detection for which later we will train the same version of YOLO with our already ready dataset of 9 GBs. OpenCV Pipelines are used to stream and get the video feed.